change settings routes and add props for icons
