+++ app/routes/drawer/routes/settings/settingsScene.js
@@ -1,15 +1,12 @@
 import React, { Component, PropTypes } from 'react';
-import { View, Text, StyleSheet } from 'react-native';
+import { View, Text, StyleSheet, Switch, TouchableOpacity, Dimensions } from 'react-native';
 import { COLOR, ThemeProvider } from 'react-native-material-ui';
+import { StackNavigator } from 'react-navigation';
 import Container from '../../container';
 import Header from '../../../../components/header.js';
-
-const uiTheme = {
-    palette: {
-        primaryColor: COLOR.grey500,
-        accentColor: COLOR.grey500
-    }
-};
+import GeneralAppInfo from './generalAppInfo/';
+import UserInfo from './userInfo/';
+const window = Dimensions.get('window');
 
 class SettingsScene extends Component {
     static propTypes = {
@@ -17,13 +14,43 @@ class SettingsScene extends Component {
     };
     render () {
         return (
-            <ThemeProvider uiTheme={uiTheme}>
+            <ThemeProvider>
                 <Container>
-                    <Header navigatorLeft={() => this.props.navigation.navigate('DrawerOpen')} title="React Settings" />
+                    <Header leftIcon="bars" navigatorLeft={() => this.props.navigation.navigate('DrawerOpen')} title="React Settings" />
                     <View style={styles.container}>
-                        <Text style={styles.welcome}>
-                          Welcome to Settings
-                        </Text>
+                        <View style={styles.touchableStyle}>
+                            <View style={styles.viewStyle}>
+                                <Text style={[
+                                    styles.textStyle, {
+                                        marginTop: 10
+                                    }
+                                ]}>
+                                  Open website in App
+                              </Text>
+                                <Switch />
+                            </View>
+                        </View>
+                        <TouchableOpacity style={styles.touchableStyle} onPress={() => this.props.navigation.navigate('UserInfo')} title="User Info">
+                            <View>
+                                <Text style={styles.textStyle}>
+                                    User Info
+                                </Text>
+                            </View>
+                        </TouchableOpacity>
+                        <TouchableOpacity style={styles.touchableStyle} onPress={() => this.props.navigation.navigate('General')} title="General">
+                            <View>
+                                <Text style={styles.textStyle}>
+                                    General App Info
+                                </Text>
+                            </View>
+                        </TouchableOpacity>
+                        <TouchableOpacity style={styles.touchableStyle} onPress={() => this.props.navigation.navigate('Login')} title="User Info">
+                            <View>
+                                <Text style={styles.textStyle}>
+                                    LogOut
+                                </Text>
+                            </View>
+                        </TouchableOpacity>
                     </View>
                 </Container>
             </ThemeProvider>
@@ -33,16 +60,44 @@ class SettingsScene extends Component {
 
 const styles = StyleSheet.create({
     container: {
-        flex: 1,
-        justifyContent: 'center',
-        alignItems: 'center',
-        backgroundColor: '#F5FCFF'
+        backgroundColor: '#FFF',
+        height: window.height
     },
-    welcome: {
+    touchableStyle: {
+        height: 50,
+        width: window.width,
+        backgroundColor: '#EEE',
+        display: 'flex',
+        justifyContent: 'center',
+        paddingLeft: 20,
+        paddingRight: 20,
+        marginTop: 10
+    },
+    viewStyle: {
+        flex: 1,
+        flexDirection: 'row',
+        justifyContent: 'space-between',
+        width: window.width - 40
+    },
+    textStyle: {
         fontSize: 20,
-        textAlign: 'center',
-        margin: 10
+        fontFamily: 'Cochin'
     }
 });
 
-export default SettingsScene;
+const SettingsStack = StackNavigator({
+    Home: {
+        screen: SettingsScene,
+        path: 'drawer'
+    },
+    General: {
+        screen: GeneralAppInfo,
+        path: 'general'
+    },
+    UserInfo: {
+        screen: UserInfo,
+        path: 'general'
+    }
+}, { headerMode: 'none' });
+
+export default SettingsStack;
