+++ app/routes/drawer/routes/tutorials/tutorialsScene.js
@@ -17,7 +17,7 @@ class TutorialsScene extends Component {
         return (
             <ThemeProvider uiTheme={uiTheme}>
                 <Container>
-                    <Header navigatorLeft={() => this.props.navigation.navigate('DrawerOpen')} title="React Tutorials" />
+                    <Header leftIcon="bars" navigatorLeft={() => this.props.navigation.navigate('DrawerOpen')} title="React Articles" />
                     <View style={styles.container}>
                         <Text style={styles.welcome}>
                             Welcome to Tutorials!
