+++ app/routes/reactWall/reactWallScene.js
@@ -20,8 +20,7 @@ class ReactWallSceen extends Component {
         return (
             <ThemeProvider uiTheme={uiTheme}>
                 <Container>
-                    <Header navigatorLeft={() => this.props.navigation.navigate('DrawerOpen')}
-                      title="React Wall" />
+                    <Header leftIcon="bars" navigatorLeft={() => this.props.navigation.navigate('DrawerOpen')} title="React Wall" />
                     <View style={styles.container}>
                         <Text style={styles.welcome}>
                           Welcome to ReactWall!
